Rename sorting module
